segunda

+++ database/migrations/2021_09_16_140912_create_semanas_table.php
@@ -15,7 +15,7 @@ class CreateSemanasTable extends Migration
     {
         Schema::create('semanas', function (Blueprint $table) {
             $table->increments('id_semana');
-            $table->double('semana');
+            $table->integer('semana');
             $table->unsignedInteger('id_partida');
